requirements.txt
ghostscript
camelot-py[cv]
pandas
pydantic
ollama
fastapi
uvicorn
python-multipart